Move type defintion to cpp.
