+++ src/QMCHamiltonians/CoulombPBCAA.h
@@ -197,25 +197,8 @@ private:
   /// Timer for offload part
   NewTimer& offload_timer_;
 
-  struct CoulombPBCAAMultiWalkerResource : public Resource
-  {
-    CoulombPBCAAMultiWalkerResource() : Resource("CoulombPBCAA") {}
-
-    std::unique_ptr<Resource> makeClone() const override
-    {
-      return std::make_unique<CoulombPBCAAMultiWalkerResource>(*this);
-    }
-
-    Vector<CoulombPBCAA::Return_t, OffloadPinnedAllocator<CoulombPBCAA::Return_t>> values_offload;
-
-    /// a walkers worth of per particle coulomb AA potential values
-    Vector<RealType> v_sample;
-
-    /// constant values per particle for coulomb AA potential
-    Vector<RealType> pp_consts;
-  };
-
   /// multiwalker shared resource
+  struct CoulombPBCAAMultiWalkerResource;
   ResourceHandle<CoulombPBCAAMultiWalkerResource> mw_res_handle_;
 
   /** constructor code factored out
